refactor: migrate web3 connection apis
