+++ packages/web3-providers/src/Connection/Flow/assets/tUSD.png
