--- packages/web3-providers/src/EVM/index.ts
@@ -1,2 +1,0 @@
-export * from './apis/Web3API.js'
-export * from './apis/Web3GasOptionAPI.js'
