--- packages/web3-providers/src/Flow/index.ts
@@ -1,1 +1,0 @@
-export * from './apis/FungibleTokenAPI.js'
