+++ packages/web3-providers/src/SmartPay/apis/FunderAPI.ts
@@ -2,7 +2,7 @@ import urlcat from 'urlcat'
 import { ChainId, type TransactionReceipt } from '@masknet/web3-shared-evm'
 import { EMPTY_LIST, type Proof } from '@masknet/shared-base'
 import { FunderAPI } from '../../types/Funder.js'
-import { Web3API } from '../../EVM/index.js'
+import { Web3API } from '../../Connection/EVM/index.js'
 import { fetchJSON } from '../../entry-helpers.js'
 import { FUNDER_PROD } from '../constants.js'
 import { BigNumber } from 'bignumber.js'
