--- packages/web3-providers/src/Solana/index.ts
@@ -1,2 +1,0 @@
-export * from './apis/FungibleTokenAPI.js'
-export * from './apis/NonFungibleTokenAPI.js'
