+++ packages/web3-providers/src/entry.ts
@@ -24,7 +24,7 @@ import { AstarAPI } from './Astar/index.js'
 import { GoPlusLabsAPI, GoPlusAuthorizationAPI } from './GoPlusLabs/index.js'
 import { NextIDProofAPI, NextIDStorageAPI } from './NextID/index.js'
 import { AlchemyEVM_API, AlchemyFlowAPI } from './Alchemy/index.js'
-import { Web3API, Web3GasOptionAPI } from './EVM/index.js'
+import { Web3API, Web3GasOptionAPI } from './Connection/EVM/index.js'
 import { MagicEdenAPI } from './MagicEden/index.js'
 import { CoinMarketCapTrendingAPI } from './CoinMarketCap/index.js'
 import { UniswapAPI } from './Uniswap/index.js'
@@ -51,7 +51,7 @@ import { NomicsAPI } from './Nomics/index.js'
 import { DSearchAPI } from './DSearch/index.js'
 import { MulticallAPI } from './Multicall/index.js'
 import { LensAPI } from './Lens/index.js'
-import { Web3SignerAPI } from './EVM/apis/Web3SignerAPI.js'
+import { Web3SignerAPI } from './Connection/EVM/apis/Web3SignerAPI.js'
 import { FlowWeb3API } from './Flow/apis/Web3API.js'
 import { SolanaWeb3API } from './Solana/apis/Web3API.js'
 import { SentryAPI } from './Sentry/index.js'
